Fix implicit conversion warnings for strlen() throughout codebase
Cast strlen() return values to appropriate types (int/long/unsigned int)
when assigning to variables or passing to functions that expect smaller
integer types. This fixes -Wshorten-64-to-32 and -Wsign-conversion
warnings during 64-bit port.

Changes:
- viola/*.c: Cast to int/long for local variables and X11 functions
- libWWW/*.c: Cast to int for string length calculations and API calls
- libXPA/src/xpa.c: Cast to int for X11 text drawing functions
- libIMG/*.c: Cast to unsigned int for lmalloc() calls

All casts are safe because string lengths in this codebase are always
within reasonable limits (URLs, filenames, UI text, etc.).

+++ src/libIMG/dither.c
@@ -69,7 +69,7 @@ Image* dither(Image* cimage, unsigned int verbose) {
     }
     image = newBitImage(cimage->width, cimage->height);
     if (cimage->title) {
-        image->title = (char*)lmalloc(strlen(cimage->title) + 12);
+        image->title = (char*)lmalloc((unsigned int)strlen(cimage->title) + 12);
         sprintf(image->title, "%s (dithered)", cimage->title);
     }
 
@@ -196,7 +196,7 @@ Image* approx(Image* cimage, unsigned int verbose) {
     }
     image = newBitImage(cimage->width, cimage->height);
     if (cimage->title) {
-        image->title = (char*)lmalloc(strlen(cimage->title) + 16);
+        image->title = (char*)lmalloc((unsigned int)strlen(cimage->title) + 16);
         sprintf(image->title, "%s (approximated)", cimage->title);
     }
 #ifdef sflkjsflkflkf
